Implement new login flow
Note: you implemented this already in teacher-features, but that was a wrong move because it's not a teacher-feature, and trying to pick out that individual commit was painful, so now you're just redoing work you've already done, which is plain annoying. If it's not a teacher feature, don't implement it in teacher-features!